Перенесен JWT Auth в Startup *.Gateway.API. Добавлен Auth класс. Иные исправления для аутентификации.
