+++ Gateway/PaymentPlatform.Gateway.API/Helpers/AppSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace PaymentPlatform.Identity.API.Helpers
+{
+    /// <summary>
+    /// Настройки в проекте Identity.
+    /// </summary>
+    public class AppSettings
+    {
+        /// <summary>
+        /// Секретный ключ.
+        /// </summary>
+        public string Secret { get; set; }
+    }
+}
