+++ Gateway/PaymentPlatform.Gateway.API/Startup.cs
@@ -1,37 +1,61 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.IdentityModel.Tokens;
 using Ocelot.DependencyInjection;
 using Ocelot.Middleware;
+using PaymentPlatform.Gateway.API.Helpers;
+using PaymentPlatform.Identity.API.Helpers;
+using System.Text;
 
 namespace PaymentPlatform.Gateway.API
 {
     public class Startup
     {
+        public IConfiguration Configuration { get; }
+
         public Startup(IHostingEnvironment env)
         {
             var builder = new ConfigurationBuilder()
                 .SetBasePath(env.ContentRootPath)
+                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
                 .AddJsonFile("configuration.json")
                 .AddEnvironmentVariables();
 
             Configuration = builder.Build();
         }
 
-        public IConfiguration Configuration { get; }
-
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddOcelot(Configuration);
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
 
-            services.AddOcelot(Configuration);
+            var appSettingSection = Configuration.GetSection("AppSettings");
+            services.Configure<AppSettings>(appSettingSection);
+
+            var appSettings = appSettingSection.Get<AppSettings>();
+            var key = Encoding.ASCII.GetBytes(appSettings.Secret);
+
+            var authOtions = new AuthOptions();
+            services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
+            .AddJwtBearer(options =>
+            {
+                //TODO: Вынести в конфиг
+                options.RequireHttpsMetadata = false;
+                options.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuer = true,
+                    ValidIssuer = authOtions.ValidIssuer,
+                    ValidateAudience = true,
+                    ValidAudience = authOtions.ValidAudience,
+                    ValidateLifetime = true,
+                    IssuerSigningKey = authOtions.GetIssuerSigningKey(),
+                    ValidateIssuerSigningKey = true,
+                };
+            });
         }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
