+++ Gateway/PaymentPlatform.Gateway.API/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "AppSettings": {
+    "Secret": "3ce1637ed40041cd94d4853d3e766c4d"
+  }
+}
